requirements.txt
django
djangorestframework
django-cors-headers
channels
daphne
requests
gunicorn
whitenoise
dj-database-url
psycopg2-binary
